Proposal for Native Histogram Text Format
Create an initial prosal for supporting native histograms in the text
format. It mainly links to the design doc for ease of comments while we
decide on a path forward. After a proposal from the design doc has some
support the rest of the markdown proposal will be filled out.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>