Proposal for Native Histogram Text Format (#32)
* Proposal for Native Histogram Text Format

Create an initial prosal for supporting native histograms in the text
format. It mainly links to the design doc for ease of comments while we
decide on a path forward. After a proposal from the design doc has some
support the rest of the markdown proposal will be filled out.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>

* Address review feedback

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>

* Update proposal with the results of the design doc

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>

* Add section explaining backwards compatability and versioning plan

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>

* Address review feedback

The significant changes are:
1. Add additional examples.
2. Add a proposal for how content negotiation will be done until OM 1.1
   is released.
3. Plural form of arrays in the schema.
4. Added support for multiple exemplars as a non-goal.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>

* Address review comments

* Tightened up some language of how to tell classic and native
  histograms apart.
* Added that whitespace is not allowed inside of a value.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>

* Add example for combined histogram with labels

Illustrate that ordering matters for this new format to aid parsing.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>

* Include support for multiple exemplars

Provide a paragraph describing support for multiple exemplars as part of
this proposal. Currently it is left ambiguous if a list of exemplars can
be included on more than native histograms, we can define that
explicitly when writing the new OpenMetrics spec.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>

* Add additional non-goals

Explicitly state that changing created timestamp semantics and allowing
custom buckets as part of the native histogram structure are not goals
for this proposal.

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>

---------

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>